Logging and return codes for Jenkins.
Summary:
Add flag for result-derived return codes.

Jenkins relies on changes-client to exit with a nonzero return code
on failure. Added to a flag so that mesos behavior is unchanged.

Add additional logging.

In particular, we now output the clientLog to stdout so that
it gets picked up by Jenkins builds.

Test Plan:
Tested locally in changes_dev. Only impacts Jenkins builds.
Tested mesos as well for good measure.

Reviewers: jukka

Reviewed By: jukka

Subscribers: mkedia

Differential Revision: https://tails.corp.dropbox.com/D112624