Add backend option, default to lvm
Summary:
OverlayFS is bad so we make this more flexible. We default to LVM
because that is what any sane setup should use.

Test Plan:
I made a deb and installed it on a mesos slave with its processes down. Manually
running a build with no snapshot (because the snapshot from changes/ is an overlayfs
one) works.

Reviewers: vishal

Reviewed By: vishal

Subscribers: mkedia, cam, jukka

Differential Revision: https://tails.corp.dropbox.com/D116002