Support snapshot artifacts for jenkins builds.
Summary:
Create a snapshot_status.json artifact reporting the snapshot status
for lxc-based jenkins-reporter builds. This allows the changes end
to receive the artifact and mark the image as active/inactive.

Test Plan:
Tested locally on changes_dev. Since this is the client there are not
noticable differences between changes_dev and prod.

Reviewers: jukka

Reviewed By: jukka

Subscribers: mkedia, vishal

Maniphest Tasks: T44631

Differential Revision: https://tails.corp.dropbox.com/D113829