Document the mesos reporter.
Summary:
In particular I focus on the characteristics of the multithreading
here as in any case with multiple threads, understanding the
interactions between the threads (in this case, the channels)
is of extreme importance as well as understanding when the methods
in this module are blocking/nonblocking.

Test Plan: Doc.

Reviewers: jukka

Reviewed By: jukka

Subscribers: mkedia, vishal

Differential Revision: https://tails.corp.dropbox.com/D117774