Fix base container creation for 'download' template.
Summary:
go-lxc bindings are lame and special case the 'download' template. It
(not lxc itself) was throwing an error for not having enough arguments
even though the resulting api call would have succeeded because we specify
the necessary arguments ourselves through ExtraArgs. In order to get around
this we are forced to special case the download template.

Test Plan:
Tested locally on a single mesos machine by installing the updated
version with dpkg -i. This is basically a revert back to what we used
to do for the 'download' template (which was broken by the change that
implemented the ability to use download-lxc)

Reviewers: armooo

Reviewed By: armooo

Subscribers: mkedia, mcot

Differential Revision: https://tails.corp.dropbox.com/D120910