Retry loading config on 404
Summary:
The race condition on the jenkins side is actually difficult to fix
because we can't create the jobstep until we have the job_data. To do
it properly we would have to commit a partially-created jobstep to
the database which may have ill consequences.

Instead we just wait for the jobstep to be created on the changes side
here with a reasonable delay of five retries with 0.250, 0.500, 1, 2,
and 4 seconds each.

Test Plan: Get a green build on changes-client and test this changes-client locally.

Reviewers: kylec, jukka

Reviewed By: jukka

Subscribers: mkedia, vishal

Differential Revision: https://tails.corp.dropbox.com/D126689