Ensure all pending Sentry reports finish before exiting
Summary:
Our individual sentry helper methods don't wait for the report to be sent, and if we
wait for pending reports before terminating the process (which is what this diff does)
they don't have to.
Also a slight refactoring so that the `defer`s can reliably fire, which they can't when we
call `os.Exit` rather than returning.

Test Plan: Not much. Suggestions?

Reviewers: anupc

Reviewed By: anupc

Subscribers: mkedia, anupc

Differential Revision: https://tails.corp.dropbox.com/D129150