Make failures not resulting for a failing exit code from a command infra failures.
Summary:
With this, the only way to get a failure result is if one of the commands exits with a non-zero value (I think).
This should give us much more accurate tracking of infra failures.
We should handle this more centrally based on the associated non-nil `error` value, but it's worth getting
accuracy in the meantime.

Test Plan: Very sparse existing unit.

Reviewers: jhance, anupc

Reviewed By: anupc

Subscribers: changesbot, mkedia, anupc

Differential Revision: https://tails.corp.dropbox.com/D129884