More rigorous Sentry logging and error reporting
Summary:
We need to know that Sentry is enabled, when it is used, and when it fails.
It is possible when calling `Capture` to listen for the `error` result (and we do)
but that requires blocking and doesn't work for the `Error` or `Message` wrappers, so
instead I've wrapped the Transport to report errors there. This won't account for
dropped errors due to full send buffers, but we'd need to have built up 100 messages for
that to happen.

Test Plan: I plan on trying it locally with a specified DSN.

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot, mkedia, anupc

Differential Revision: https://tails.corp.dropbox.com/D134164