Make client.Log close-safe
Summary:
We 'time out' certain operations, allowing them to proceed in the background and potentially continue logging.
As such, we don't really have centralized knowledge of when the logger will no longer be written to, so we can't
close the channel without risking send-after-close (which is always an error).
Thus, we never close the channel, and instead make the 'Close' method ensure that sends will not block or cause
problems. In the process, this hides direct access to the channel.

Test Plan: Unit

Reviewers: nate, anupc

Reviewed By: anupc

Subscribers: changesbot, mkedia

Differential Revision: https://tails.corp.dropbox.com/D135872