Set all lxc configs in the config file.
Summary:
`SetCgroupItem` seems to only work after a container is started.
We can set the cpu and mem limits after starting the container
but I don't see a reason to not just do it in the config file
like the other settings.

Test Plan:
I was able to repro the problem in my dev vm. Setting the configs
after the container starts does work.
This seems to work too. Verified that it's not doing nothing by
setting bad values for them and seeing the container fail to
start correctly.

Reviewers: kylec, anupc

Reviewed By: anupc

Subscribers: changesbot, mkedia, jukka

Differential Revision: https://tails.corp.dropbox.com/D137780