Generate lxc-free deb
Summary:
We can't yet put LXC on all of our slaves, but we should use changes-client everywhere.
This adds a tag 'nolxc' that results in a no-lxc build, and generates a deb with it that
also provides 'changes-client' and doesn't require LXC.
I kept the LXC flags for convenience (bad adapter choice already a better failure than
missing flag).
I added a new build tag because I'm pretty sure the 'lxc' build tag is useless to us,
but I didn't confirm that enough to remove it.

Test Plan: Minimal thus far.

Reviewers: anupc

Reviewed By: anupc

Subscribers: zbarsky, changesbot, mkedia, nate, paulruan

Differential Revision: https://tails.corp.dropbox.com/D140822