Improve error message from GetDebugConfig
Summary:
It should be possible to just log the error and move on, and for that it needs
to mention where it is from and for what key.

Test Plan:
Existing unit; could add a test that the error string contains the key or something,
but that seemed overkill for something we don't depend on.

Reviewers: nate, paulruan

Reviewed By: paulruan

Subscribers: changesbot, mkedia, anupc

Differential Revision: https://tails.corp.dropbox.com/D141894