Add debug config flag to keep container
Summary:
Sometimes we want to inspect a container, but we can't always do so by fiddling with flags or intra-repo scripts.
This allows us to tell changes-client to do it directly via debugConfig, and is an endtime rather than a bool because
unintentionally leaving it enabled is terrible and it shouldn't be easy to do that.

Test Plan: Unit

Reviewers: nate

Reviewed By: nate

Subscribers: changesbot, mkedia, anupc, josiah

Differential Revision: https://tails.corp.dropbox.com/D144142