Parse the intended snapshot ID from the jobstep config and verify it.
Summary:
We get the destination snapshot ID from both a flag and the config. We
don't use the config, but we should, and this change starts us on that
path by parsing it and complaining when it doesn't match the flag value.
As far as I can tell, the logic is the same in both cases, but erring on
the side of caution, since being wrong here could break a lot of things.

The next steps will be:
  1. Use the config value instead of the flag value.
  2. Stop providing the flag.
  3. Remove the flag.

Test Plan:
Some unit tests. Arguably overkill, but makes the later steps easy
changes with easy testing.

Reviewers: nate

Reviewed By: nate

Subscribers: changesbot, mkedia, anupc

Differential Revision: https://tails.corp.dropbox.com/D145126