Update client.Log to flush after 10k or 4 seconds
Summary:
This'll make the live log view a little less responsive, but will potentially
significantly reduce load on artifactstore.
Note that the limits seem to only apply to piped log data, not direct calls to Printf,
but my assumption is that the vast majority of log data is piped.

Test Plan: None

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot, nate

Differential Revision: https://tails.corp.dropbox.com/D149092