Make it possible to have two filelogs with the same name
Summary:
The local path of a filelog isn't particularly important, and having a stable path that
can't be reused causes errors when repeating runs for testing or in a joyride setup.
This moves to a randomly chosen jobstep directory to preserve the existing structure but avoid
conflicts.
In the future when we have more filelogs, we'll likely want to share a single unique directory for
all filelogs in a run for simplicity, but I haven't done that here because it isn't important or yet
useful.

Test Plan: Unit

Reviewers: nate, anupc

Reviewed By: anupc

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D161196