Consider it an infra failure if an infra_setup command fails
Summary:
This will make it clear that clone/patch application failures (and later
blacklisting issues) are our problem.
Ideally, this might be a flag on Command (I have a diff for that), but
'infra_setup' covers all commands we should infra fail currently, and
I think any 'infra_setup' should result in an infra failure, so this
is a faster, easier, and currently complete fix.

Test Plan: Iffy unit

Reviewers: nate

Reviewed By: nate

Subscribers: changesbot, anupc

Differential Revision: https://tails.corp.dropbox.com/D173561