Add realistic fnMatch benchmark
Summary:
This takes a somewhat mangled version of a large live blacklist and runs it on a small set of files repeatedly.
It'd be more realistic if the set of files were huge, but I think this is good enough to be useful.
This already helped me determine that an algorithmic improvement in chunk matching actually makes things worse.

Test Plan: Manual

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D178387